move scripts to scripts folder; update package.json run scripts move generator to scripts lint & autofix newly generated documentation.ts
