+++ package.json
@@ -171,8 +171,8 @@
     },
     "scripts": {
         "vscode:prepublish": "npm run updateDependencies && npm run package",
-        "updateDependencies": "node updateDependencies.js",
-        "generateDocumentation": "node src/language/documentation/generator.js",
+        "updateDependencies": "node scripts/updateDependencies.js",
+        "generateDocumentation": "node scripts/language/documentation/generator.js && eslint src/language/documentation/documentation.ts --fix",
         "postinstall": "npm run updateDependencies",
         "compile": "webpack",
         "watch": "npm run generateDocumentation && webpack --watch",
